Revert "Проект 6. Доработан класс InMemoryHistoryManager и добавлены тесты"
