--- src/Test/model/TaskTest.java
@@ -1,21 +1,0 @@
-package Test.model;
-
-import com.yandex.tracker.model.Task;
-import com.yandex.tracker.service.TaskStatus;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class TaskTest {
-
-    @Test
-    void testTaskEqualityById() {
-        Task task1 = new Task("Task 1", "Description 1", TaskStatus.NEW);
-        Task task2 = new Task("Task 2", "Description 2", TaskStatus.NEW);
-
-        task1.setId(1);
-        task2.setId(1);
-
-        assertEquals(task1, task2, "Задачи должны быть равны по id.");
-    }
-}
