+++ src/com/yandex/tracker/service/HistoryManager.java
@@ -4,7 +4,7 @@ import com.yandex.tracker.model.Task;
 import java.util.List;
 
 public interface HistoryManager {
+
    void add(Task task);
-   void remove(int id);
    List<Task> getHistory();
 }
