+++ src/com/yandex/tracker/service/Managers.java
@@ -5,7 +5,7 @@ public class Managers {
     private Managers() {
     }
     public static TaskManager getInMemoryTaskManager(HistoryManager historyManager) {
-        return new InMemoryTaskManager(historyManager);
+        return new InMemoryTaskManager(historyManager); // Возвращаем конкретную реализацию TaskManager
     }
     public static HistoryManager getDefaultHistory(){
         return new InMemoryHistoryManager();
